move useLoginFlow to separate file
