+++ apps/web/src/components/Shared/Follow.tsx
@@ -1,6 +1,7 @@
 import type { ApolloCache } from '@apollo/client';
 import { Button } from '@components/UI/Button';
 import { Spinner } from '@components/UI/Spinner';
+import useLoginFlow from '@components/utils/hooks/useLoginFlow';
 import { UserAddIcon } from '@heroicons/react/outline';
 import { Analytics } from '@lib/analytics';
 import getSignature from '@lib/getSignature';
@@ -17,8 +18,6 @@ import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { PROFILE } from 'src/tracking';
 import { useAccount, useContractWrite, useSignTypedData } from 'wagmi';
-
-import { useLoginFlow } from './GlobalModals';
 
 export enum FollowSource {
   WHO_TO_FOLLOW = 'who_to_follow',
